SpreadsheetConverterContextTesting2 extracted SpreadsheetConverterContextTesting
