+++ src/main/java/walkingkooka/spreadsheet/format/SpreadsheetFormatterContextTesting2.java
@@ -18,7 +18,7 @@
 package walkingkooka.spreadsheet.format;
 
 import org.junit.jupiter.api.Test;
-import walkingkooka.spreadsheet.convert.SpreadsheetConverterContextTesting;
+import walkingkooka.spreadsheet.convert.SpreadsheetConverterContextTesting2;
 import walkingkooka.tree.expression.convert.ExpressionNumberConverterContextTesting2;
 import walkingkooka.tree.text.TextNode;
 
@@ -27,7 +27,7 @@ import java.util.Optional;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public interface SpreadsheetFormatterContextTesting2<C extends SpreadsheetFormatterContext> extends SpreadsheetFormatterContextTesting<C>,
-    SpreadsheetConverterContextTesting<C>,
+    SpreadsheetConverterContextTesting2<C>,
     ExpressionNumberConverterContextTesting2<C> {
 
     @Override
